Include the update parser within the updater class as there is no reason for it to be in it's own class
